+++ scripts/install-hooks.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+CWD="$(cd "$(dirname "$0")" && pwd)"
+
+cp --no-target-directory --recursive --verbose "$CWD/hooks" "$CWD/../.git/hooks"